test: optimize test_algorithm
